+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Several `--int64`, `--float32` related fixes
 - Many other minor bugs uncovered (and fixed) by Claude
 - Fixed several C++ compilation warnings
+- Hash values are now same type as shedskin int type
 
 ### Optimized